requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
colorama==0.4.5
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
